refactor: split records and nameservers tests by CRUD operation
Further split the records (1024 lines) and nameservers (668 lines) test
files into smaller files organized by CRUD operation: create, read,
update, delete. Each file is now 120-430 lines. Inlined the table-driven
error tests back into their respective operation files for better
locality.
